+++ static_report/src/pages/SRTablesListPage.tsx
@@ -16,7 +16,7 @@ import { useDocumentTitle } from '../hooks/useDocumentTitle';
 import { SaferSRSchema } from '../types';
 import { TableListItem } from '../components/shared/Tables/TableList/TableListItem';
 import { tableListGridTempCols, tableListWidth } from '../utils/layout';
-import { useReportStore } from '../components/shared/Tables/store';
+import { useReportStore } from '../utils/store';
 import { TableColumnSchemaList } from '../components/shared/Tables/TableList/TableColumnSchemaList';
 
 type Props = { data: SaferSRSchema };
